NormalBitSequence(99.99%) + debug/renamings

+++ Sequence.h
@@ -7,7 +7,7 @@ template <class T> class Sequence{
         virtual T get_first() const = 0;
         virtual T get_last() const = 0;
         virtual T get(int index) const = 0;
-        virtual int get_length() const = 0;
+        virtual size_t get_length() const = 0;
         virtual Sequence<T> *append(T item) = 0;
         virtual Sequence<T> *prepend(T item) = 0;
         virtual Sequence<T> *insert_at(T item, int index) = 0;
